вынести появление и уничтожение врагов в отдельный класс, вынести обновление волн в отдельный класс
